requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
addict=2.4.0=pypi_0
ansi2html=1.9.1=pypi_0
attrs=23.2.0=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.12.12=hecd8cb5_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
dash=2.14.2=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
fastjsonschema=2.19.1=pypi_0
flask=3.0.0=pypi_0
fonttools=4.47.0=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.1=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.1=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hecd8cb5_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.8=pypi_0
numpy=1.26.3=pypi_0
open3d=0.18.0=pypi_0
openssl=3.0.12=hca72f7f_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py311hecd8cb5_0
platformdirs=4.1.0=pypi_0
plotly=5.18.0=pypi_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.11.7=hf27a42d_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.32.1=pypi_0
requests=2.31.0=pypi_0
retrying=1.3.4=pypi_0
rpds-py=0.16.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
setuptools=68.2.2=py311hecd8cb5_0
six=1.16.0=pypi_0
sqlite=3.41.2=h6c40b1e_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h5d9f67b_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
urllib3=2.1.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311hecd8cb5_0
xz=5.4.5=h6c40b1e_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h4dc903c_0
